removed unnecessary codes
