+++ microservices/userManagement/pom.xml
@@ -33,6 +33,13 @@
 		    <version>8.0.26</version>
 		</dependency>
 
+	    <dependency>
+	        <groupId>org.projectlombok</groupId>
+	        <artifactId>lombok</artifactId>
+	        <version>1.18.22</version> <!-- Use the latest version -->
+	        <scope>provided</scope>
+	    </dependency>
+    
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-test</artifactId>
