+++ microservices/userManagement/src/main/java/com/healthcare/appointmentScheduler/userManagement/entity/Doctor.java
