--- microservices/userManagement/userManagement/src/main/java/com/healthcare/appointmentScheduler/userManagement/UserManagementApplication.java
@@ -1,14 +1,0 @@
-package com.healthcare.appointmentScheduler.userManagement;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.domain.EntityScan;
-
-
-@SpringBootApplication
-@EntityScan(basePackages = "com.healthcare.appointmentScheduler.userManagement.entity")
-public class UserManagementApplication {
-	public static void main(String[] args) {
-		SpringApplication.run(UserManagementApplication.class, args);
-	}
-}
