--- microservices/userManagement/userManagement/src/main/java/com/healthcare/appointmentScheduler/userManagement/repository/UserRepository.java
@@ -1,11 +1,0 @@
-package com.healthcare.appointmentScheduler.userManagement.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.healthcare.appointmentScheduler.userManagement.entity.User;
-
-public interface UserRepository extends JpaRepository<User, Integer>{
-
-	
-	
-}
