--- microservices/userManagement/userManagement/src/main/resources/application.properties
@@ -1,8 +1,0 @@
-server.port=8081
-spring.datasource.url=jdbc:mysql://localhost:3306/userDB?useSSL=false
-spring.datasource.username=root
-spring.datasource.password=password
-spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
-spring.jpa.properties.hibernate.dilect=org.hibernate.dialect.MySQL8Dialect
-spring.jpa.hibernate.auto-ddl=update
-spring.jpa.show-sql=true
