--- microservices/userManagement/userManagement/src/test/java/com/healthcare/appointmentScheduler/userManagement/UserManagementApplicationTests.java
@@ -1,13 +1,0 @@
-package com.healthcare.appointmentScheduler.userManagement;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class UserManagementApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
